refactor: extract OptionRowTitleWithStatus from OptionRowLHN
